add tests for modified reporters
