+++ lib/log-processor/log-processor.mjs
@@ -209,7 +209,11 @@ export default class LogProcessor {
 
     const reporter = new this.reporter(this.stream, overallInfo);
 
-    return await reporter.report(this.options);
+    const result = await reporter.report(this.options);
+
+    return new Promise(resolve => {
+      this.stream.end(() => resolve(result));
+    });
   }
 
   cloneSummaries() {
