+++ lib/log-processor/reporters/base-reporter.cjs
@@ -19,6 +19,7 @@ class BaseReporter {
 
     this.runSummaries = this.overallInfo.runSummaries;
     this.routeMetricsResults = this.overallInfo.routeMetricsResults;
+    this.isStdout = stream.fd === process.stdout.fd;
   }
 
   /**
@@ -30,7 +31,7 @@ class BaseReporter {
     const {file, linesRead, charCount} = this.overallInfo;
     const yOrIes = count === 1 ? 'summary' : 'summaries';
     // eslint-disable-next-line no-console
-    console.log(`[[read ${count} ${yOrIes} (${linesRead} lines, ${charCount} chars) from ${file}]]\n`);
+    console.log(`[[read ${count} ${yOrIes} (${linesRead} lines, ${charCount} chars) from ${file}]]`);
   }
 
   report() {
