+++ lib/log-processor/reporters/csv.cjs
@@ -1,6 +1,7 @@
 'use strict';
 
 const BaseReporter = require('./base-reporter.cjs');
+const {f2, iso, stats, percentile} = require('./base-reporter.cjs');
 
 class Reporter extends BaseReporter {
   constructor(stream, overallInfo, options) {
@@ -12,51 +13,75 @@ class Reporter extends BaseReporter {
     // output informational text
     this.reportOverall();
 
-    for (const s of this.runSummaries) {
+    for (const rmr of this.routeMetricsResults) {
       /* eslint-disable no-console */
-      console.log(`[start ${iso(s.firstTimestamp)}, end ${iso(s.lastTimestamp)}]`);
-      const rc = s.routes.length;
+      //this.stream.write(`[start ${iso(rmr.earliestTimestamp)}, end ${iso(rmr.latestTimestamp)}]\n`);
+      // if (!this.isStdout) {
+      console.log(`[start ${iso(rmr.earliestTimestamp)}, end ${iso(rmr.latestTimestamp)}]`);
+      // }
+      // this is routes, while what is calculated after this is route/status
+      // combinations. not sure that is what should be reported.
+      //const routes = Object.entries(rmr.routesIndex);
+
+      const groups = rmr.route.groupBy(options.template.routes || []);
+      //const compat = rmr.route.compatGetRoutesByStatusCode();
+      const rc = groups.size;
+
       let totalObs = 0;
-      for (const [, timesByStatus] of s.routes) {
-        for (const status in timesByStatus) {
-          totalObs += timesByStatus[status].length;
+      for (const statusCodes of groups.values()) {
+        for (const status in statusCodes) {
+          totalObs += statusCodes[status].length;
           if (!options.microseconds) {
-            timesByStatus[status] = timesByStatus[status].map(t => Math.round(t / 1000));
+            statusCodes[status] = statusCodes[status].map(t => Math.round(t / 1000));
           }
         }
       }
+      //this.stream.write(`[total time measurements ${totalObs} across ${rc} routes]\n`);
+      // if (!this.isStdout) {
       console.log(`[total time measurements ${totalObs} across ${rc} routes]`);
+      // }
 
-      // note the agent is loaded if it is
-      for (const p of s.patches) {
-        if (!p.entry.name.startsWith('http')) {
+      for (const p of rmr.patch.logObjects) {
+        if (p.entry.name.startsWith('@contrast/agent')) {
+          // this.stream.write(`[${p.entry.name} loaded]\n`);
+          // if (!this.isStdout) {
           console.log(`[${p.entry.name} loaded]`);
+          // }
           break;
         }
       }
 
-      // and output time-series info
-      const count = this.getTimeSeriesCount(s.timeSeries);
-      if (count) {
-        let gcInfo = '';
-        if (s.timeSeries.gc.count !== 0) {
-          gcInfo = ` (gc count ${s.timeSeries.gc.gcCount} gc time ${s.timeSeries.gc.gcTime})`;
-        }
-        console.log(`[time-series lines processed ${count}${gcInfo}]`);
+      let totalTimeSeries = 0;
+      for (const ts of ['proc', 'gc', 'eventloop']) {
+        totalTimeSeries += rmr[ts].count;
       }
-      /* eslint-enable no-console */
-
-      // sort the routes so the percentiles work correctly.
-      s.routes.sort((a, b) => a === b ? 0 : a[0] < b[0] ? -1 : 1);
+      let gcInfo = '';
+      if (totalTimeSeries) {
+        if (rmr.gc.count !== 0) {
+          gcInfo = ` (gc count ${rmr.gc.totalCount} gc time ${rmr.gc.totalTime})`;
+        }
+      }
+      //this.stream.write(`[time-series lines processed ${totalTimeSeries}${gcInfo}]\n`);
+      // if (!this.isStdout) {
+      console.log(`[time-series lines processed ${totalTimeSeries}${gcInfo}]`);
+      // }
 
       const percentiles = this.percentiles.join(', ');
-
       this.stream.write(`route, status, n, mean, stddev, percentiles: ${percentiles}\n`);
-      // route, status, n, ...percentiles
-      if (!options.template) {
-        this.writeSimple(s);
-      } else {
-        this.writeTemplateDriven(s, options.template);
+      const sorted = Array.from(groups.keys());
+      sorted.sort();
+
+      for (const route of sorted) {
+        const statusCodes = groups.get(route);
+        for (const status in statusCodes) {
+          const times = statusCodes[status];
+          const {n, mean, stddev} = stats(times);
+          const percentiles = percentile(this.percentiles, times);
+          const line = [
+            route, status, n, f2(mean), f2(stddev), ...percentiles
+          ].join(',');
+          this.stream.write(`${line}\n`);
+        }
       }
       // console.log is synchronous and stream writes are not. informational text
       // is always written to the console. if the report is also being written to
@@ -67,155 +92,6 @@ class Reporter extends BaseReporter {
       });
     }
   }
-
-  /**
-   * Write the summary based on the raw route names.
-   *
-   * @param {object} summary a run summary.
-   */
-  writeSimple(summary) {
-    const {routes} = summary;
-
-    // for each collected metric
-    for (const [path, timesByStatus] of routes) {
-      // for each status code
-      for (const status in timesByStatus) {
-        const lowToHigh = (a, b) => a === b ? 0 : a < b ? -1 : 1;
-        const times = timesByStatus[status].sort(lowToHigh);
-        const {n, mean, stddev} = stats(times);
-        const line = [
-          path, status, n, f2(mean), f2(stddev), ...percentile(this.percentiles, times)
-        ].join(',');
-        //waitWriter(`${line}\n`);
-        this.stream.write(`${line}\n`);
-      }
-    }
-  }
-
-
-  /**
-   * Write the summary based grouped into buckets defined by the template.
-   *
-   * @param {object} summary a run summary
-   * @param {object} template a template definition
-   */
-  writeTemplateDriven(summary, template) {
-    // create config-defined buckets
-    const buckets = {};
-    const rawBuckets = {};
-    const {routes, meta} = summary;
-
-    // for each collected metric
-    for (const [key, timesByStatus] of routes) {
-      // get the already parsed items
-      const {method, path} = meta.keyToProperties[key];
-      // now for each config-defined bucket
-      let matched = false;
-      for (const bucket of template.routes) {
-        if (bucket.method === method) {
-          if (bucket.startsWith && path.startsWith(bucket.startsWith)) {
-            // store in bucket
-            this.saveMatch(buckets, bucket.name, timesByStatus);
-            matched = true;
-          } else if (bucket.regex) {
-            const m = path.match(bucket.regex);
-            if (m) {
-              // store in bucket
-              this.saveMatch(buckets, bucket.name, timesByStatus);
-              matched = true;
-            } else if (bucket.pattern && path === bucket.pattern) {
-              this.saveMatch(buckets, bucket.name, timesByStatus);
-              matched = true;
-            } else {
-              // hmm. either at least one match type must be present. and that
-              // should have been verified already.
-            }
-          }
-        }
-      }
-      if (!matched) {
-        // it didn't match any of the template-defined route buckets.
-        // put it in its own bucket, like default behavior
-        this.saveMatch(rawBuckets, key, timesByStatus);
-      }
-    }
-
-    // order keys so the template-named buckets come before the default-named
-    // buckets (which are really just the raw routes).
-    for (const key in rawBuckets) {
-      buckets[key] = rawBuckets[key];
-    }
-
-    for (const bucketName in buckets) {
-      const timesByStatus = buckets[bucketName];
-      for (const status in timesByStatus) {
-        const lowToHigh = (a, b) => a === b ? 0 : a < b ? -1 : 1;
-        const times = timesByStatus[status].sort(lowToHigh);
-        const {n, mean, stddev} = stats(times);
-        const percentiles = percentile(this.percentiles, times);
-        const line = [
-          bucketName, status, n, f2(mean), f2(stddev), ...percentiles
-        ].join(',');
-        this.stream.write(`${line}\n`);
-      }
-    }
-  }
-
-  saveMatch(buckets, name, timesByStatus) {
-    let bucket = buckets[name];
-    if (!bucket) {
-      bucket = buckets[name] = {};
-    }
-    for (const status in timesByStatus) {
-      if (!bucket[status]) {
-        bucket[status] = timesByStatus[status].slice();
-      } else {
-        bucket[status] = bucket[status].concat(timesByStatus[status]);
-      }
-    }
-  }
-}
-
-//
-// formatting helpers
-//
-function f2(n) {
-  return n.toFixed(2);
-}
-
-function iso(ts) {
-  return new Date(ts).toISOString();
-}
-
-//
-// summarization functions
-//
-function stats(array) {
-  const n = array.length;
-  const total = array.reduce((tot, v) => tot + v, 0);
-  const mean = total / n;
-  const stddev = variance(array) ** 0.5;
-
-  return {n, total, mean, stddev};
-}
-
-function mean(array) {
-  return array.reduce((tot, v) => tot + v, 0) / array.length;
-}
-
-function variance(array) {
-  const average = mean(array);
-  return mean(array.map((num) => (num - average) ** 2));
-}
-
-
-function percentile(percentiles, array) {
-  return percentiles.map(p => pctile(p, array));
-}
-
-function pctile(p, list) {
-  if (p === 0) return list[0];
-  return list[Math.ceil(list.length * p) - 1];
 }
 
 module.exports = Reporter;
