+++ lib/log-processor/reporters/json-deprecated.cjs
@@ -12,26 +12,19 @@ class Reporter extends BaseReporter {
   }
 
   async report() {
-    // don't output lines in the JSON summary
+    // don't output lines in the JSON summary. they were added so that
+    // the log-processor/utils/ could replace the legacy code.
     const logLines = this.runSummaries.map(s => {
       const lines = s.lines;
       delete s.lines;
       return lines;
     });
     const summaries = JSON.stringify(this.runSummaries, null, 2);
-
-    const needToWait = false === this.stream.write(summaries);
-
-    if (needToWait) {
-      await new Promise(resolve => {
-        this.stream.once('drain', resolve);
-      });
-    }
+    this.stream.write(summaries);
     // restore the lines
     for (let i = 0; i < this.runSummaries.length; i++) {
       this.runSummaries[i].lines = logLines[i];
     }
-
     return new Promise(resolve => {
       this.stream.write('\n', resolve);
     });
