+++ lib/log-processor/utils/index.mjs
@@ -56,7 +56,15 @@ export default class RouteMetricsResults {
   //
   preprocess() {
     for (const logObject of this.logObjects) {
-      // handle routes differently than all the others.
+      // routes get indexed by a key that depends elements of the url and
+      // whether the user specifies fold rules. fold rules are used to merge
+      // multiple urls into a single name, e.g., /api/users/1, /api/users/2.
+      //
+      // routes are also stored like any other type, but no other types are
+      // handled multiple ways like routes.
+      //
+      // TODO NODE-3712 currently implemented in route-metrics log-processor
+      // but should be here
       if (logObject.type === 'route') {
         let url = logObject.entry.url;
         for (const foldRule of this.foldRules) {
@@ -72,8 +80,7 @@ export default class RouteMetricsResults {
           this.routesIndex[routeKey] = route;
         }
         route.add(logObject);
-      }
-      else if (logObject.type === 'unknown-config-items') {
+      } else if (logObject.type === 'unknown-config-items') {
         continue;
       }
